zk-kyc-agent/requirements.txt
langchain>=0.3
langchain-core>=0.3
langchain-community>=0.3
langchain-openai>=0.2
pydantic>=2.7
python-dotenv>=1.0
requests>=2.32